requirements.txt
# Telegram
python-telegram-bot==21.9

# LangChain + OpenAI
langchain==0.3.14
langchain-openai==0.2.14
langchain-core==0.3.29
openai==1.59.6

# Google APIs
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Scheduling
APScheduler==3.10.4

# HTTP (async)
httpx==0.27.2

# Utils
python-dotenv==1.0.1
tenacity==8.5.0
mutagen==1.47.0
